pyproject.toml
[tool.poetry]
name = "ui_api_tests_para_bank_reqres"
version = "0.1.0"
description = ""
authors = ["Mochinut <waka332@Mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
selene = "2.0.0b5"
allure-pytest = "^2.13.2"
pytest = "^7.3.2"
Faker = "^19.2.0"
jsonschema = "^4.18.0"
curlify = "^2.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
